vinyl: trigger dump before hitting the memory limit
Hitting the global memory limit results in throttling transactions, so
we should start dumping ranges beforehand. This patch introduces the
memory barrier, which when exceeded triggers dump. The gap between the
barrier and the limit is set to such a value that should allow us to
complete dumping the biggest range by the time the limit is hit,
provided insertions will continue to come at the same rate they have
been coming recently, i.e.

  barrier = limit - range_size * insertion_rate / dump_bandwidth

After this change we can probably get rid of the range size limit.